scout/requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
python-dotenv
python-jose[cryptography]
redis
aioredis
